Reorganized renderer tests
