+++ tests/plotting/matplotlib/testrenderer.py
@@ -1,0 +1,63 @@
+from unittest import SkipTest
+from nose.plugins.attrib import attr
+
+import numpy as np
+
+from holoviews import HoloMap, Image, ItemTable, Store, GridSpace, Table
+from holoviews.element.comparison import ComparisonTestCase
+
+try:
+    from holoviews.plotting.mpl import MPLRenderer
+except:
+    pass
+
+
+@attr(optional=1)
+class MPLRendererTest(ComparisonTestCase):
+    """
+    Note if not possible to compare the hashes of SVG and WebM formats
+    as the hashes are not stable across exports.
+    """
+
+    def setUp(self):
+        if 'matplotlib' not in Store.renderers:
+            raise SkipTest("Matplotlib required to test widgets")
+
+        self.basename = 'no-file'
+        self.image1 = Image(np.array([[0,1],[2,3]]), label='Image1')
+        self.image2 = Image(np.array([[1,0],[4,-2]]), label='Image2')
+        self.map1 = HoloMap({1:self.image1, 2:self.image2}, label='TestMap')
+
+        self.unicode_table = ItemTable([('β','Δ1'), ('°C', '3×4')],
+                                       label='Poincaré', group='α Festkörperphysik')
+
+        self.renderer = MPLRenderer.instance()
+
+    def test_get_size_single_plot(self):
+        plot = self.renderer.get_plot(self.image1)
+        w, h = self.renderer.get_size(plot)
+        self.assertEqual((w, h), (288, 288))
+
+    def test_get_size_row_plot(self):
+        plot = self.renderer.get_plot(self.image1+self.image2)
+        w, h = self.renderer.get_size(plot)
+        self.assertEqual((w, h), (576, 255))
+
+    def test_get_size_column_plot(self):
+        plot = self.renderer.get_plot((self.image1+self.image2).cols(1))
+        w, h = self.renderer.get_size(plot)
+        self.assertEqual((w, h), (288, 505))
+
+    def test_get_size_grid_plot(self):
+        grid = GridSpace({(i, j): self.image1 for i in range(3) for j in range(3)})
+        plot = self.renderer.get_plot(grid)
+        w, h = self.renderer.get_size(plot)
+        self.assertEqual((w, h), (345, 345))
+
+    def test_get_size_table(self):
+        table = Table(range(10), kdims=['x'])
+        plot = self.renderer.get_plot(table)
+        w, h = self.renderer.get_size(plot)
+        self.assertEqual((w, h), (288, 288))
+
+
